Ensure segment_identity_daily step tables are created
